refactor: [#639] Tracker Checker: extract mod for Health checks
