+++ src/console/clients/checker/service.rs
@@ -1,9 +1,7 @@
 use std::net::SocketAddr;
 use std::sync::Arc;
-use std::time::Duration;
 
-use colored::Colorize;
-use reqwest::{Client as HttpClient, Url};
+use reqwest::Url;
 
 use super::checks;
 use super::config::Configuration;
@@ -37,52 +35,8 @@ impl Service {
 
         checks::http::run(&self.config.http_trackers, &self.console, &mut check_results).await;
 
-        self.run_health_checks(&mut check_results).await;
+        checks::health::run(&self.config.health_checks, &self.console, &mut check_results).await;
 
         check_results
-    }
-
-    async fn run_health_checks(&self, check_results: &mut Vec<CheckResult>) {
-        self.console.println("Health checks ...");
-
-        for health_check_url in &self.config.health_checks {
-            match self.run_health_check(health_check_url.clone()).await {
-                Ok(()) => check_results.push(Ok(())),
-                Err(err) => check_results.push(Err(err)),
-            }
-        }
-    }
-
-    async fn run_health_check(&self, url: Url) -> Result<(), CheckError> {
-        let client = HttpClient::builder().timeout(Duration::from_secs(5)).build().unwrap();
-
-        let colored_url = url.to_string().yellow();
-
-        match client.get(url.clone()).send().await {
-            Ok(response) => {
-                if response.status().is_success() {
-                    self.console
-                        .println(&format!("{} - Health API at {} is OK", "✓".green(), colored_url));
-                    Ok(())
-                } else {
-                    self.console.eprintln(&format!(
-                        "{} - Health API at {} is failing: {:?}",
-                        "✗".red(),
-                        colored_url,
-                        response
-                    ));
-                    Err(CheckError::HealthCheckError { url })
-                }
-            }
-            Err(err) => {
-                self.console.eprintln(&format!(
-                    "{} - Health API at {} is failing: {:?}",
-                    "✗".red(),
-                    colored_url,
-                    err
-                ));
-                Err(CheckError::HealthCheckError { url })
-            }
-        }
     }
 }
